(piq) Make :piqtype directive optional in .piq files
There was an artificial restriction that :piqtype directive (specifying default
object type) should always be present in .piq (and .wire) files.

This change removes the restriction and allows the default object type to be
specified as "--piqtype" command-line parameter for "piqi convert" and "piqi
check".

When Piqi is used as OCaml or Erlang data serialization system, types for loaded
Piq objects are known up-front and explicit :piqtype directive is not necessary.

Overall, always enforcing :piqtype presense was reducing flexibility and causing
unnecessary redundancy.