(piqi) Add a new "piqi convert --gen-extended-piqi-any" flag
This flag affects piqi-any values represented in JSON and XML formats. When
specified, an extended version of piqi-any representation is used in the
conversion result. In addition to the original JSON or XML value, it includes
Piqi type name (if known), Protobuf representation (if known or can be derived),
and a special marker indicating that this is an extended piqi-any
representation.

For example, this flag changes relavent portion of
"piqi convert -t json piqi.piqi" output from

   "default": "required",

to

   "default": {
      "piqi_type": "piqi-any",
      "type": "piqi/field-mode",
      "protobuf": "CN+iipMB",
      "json": "required"
    },

This feature is necessary for future Piqi compilers implemented in various
languages that take portable JSON or XML .piqi representation as input.