piqi compile: remove --normalize-names flag
The original assumption was that by having this implemented in "piqi compile",
piqi compiler backends won't need to implement this logic. It turned out not to
be true, because altering names of the original spec affects multi-format
serialization. We don't want that. So, there's no other choice but have this
logic implemented in all piqi compiler backends directly.