adds commands to piqi frontend (#70)
When testing installation using opam, I noticed that `piqi` command
doesn't list any commands. Indeed, the `run` module doesn't register
any and the commands are provided by other modules that are no
referenced in run. This commit fixes the issue and packs the commands
implementation into the private `commands` library, which is linked
with `-linkall` to the main executable.

I made the library private by default, but we can easily make it
public, if necessary.